adds 2 minting functions
First one: normal minting function that supports batch mints, we can use this if we decide to add a price. If we decide for a price, then minting multiple tokens is more convenient for the users.

Second one: the minting from the MA contract, with a verified hashed message. This could be handy if we want to have a allowlist. We could also use this if we want to limit the number of mints per wallet again.